GgufRunner/pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "llama-cpp-python>=0.3.8",
    "numpy>=2.2.4",
    "streamlit>=1.43.2",
]
